# lib/api-zod/src/generated/types/companyDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CapitalBridge API
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from "./company";
import type { CompanyDetailDocumentsItem } from "./companyDetailDocumentsItem";
import type { CompanyDetailFinancialsItem } from "./companyDetailFinancialsItem";
import type { CompanyDetailTeamItem } from "./companyDetailTeamItem";
import type { CompanyDetailUseOfFundsItem } from "./companyDetailUseOfFundsItem";

export type CompanyDetail = Company & {
  overview: string;
  useOfFunds: CompanyDetailUseOfFundsItem[];
  highlights: string[];
  financials: CompanyDetailFinancialsItem[];
  team: CompanyDetailTeamItem[];
  documents: CompanyDetailDocumentsItem[];
};