# lib/api-zod/src/generated/types/createCompanyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CapitalBridge API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCompanyInputStage } from "./createCompanyInputStage";

export interface CreateCompanyInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 4 */
  tagline: string;
  industry: string;
  stage: CreateCompanyInputStage;
  headquarters: string;
  foundedYear: number;
  annualRevenue: number;
  ebitdaMargin?: number;
  growthRate?: number;
  raiseAmount: number;
  valuation: number;
  minimumInvestment: number;
  /** @minLength 20 */
  overview: string;
}