# lib/api-zod/src/generated/types/createInvestorInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CapitalBridge API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInvestorInputType } from "./createInvestorInputType";

export interface CreateInvestorInput {
  /** @minLength 2 */
  name: string;
  type: CreateInvestorInputType;
  tagline?: string;
  location: string;
  aum?: number;
  checkSizeMin: number;
  checkSizeMax: number;
  focusSectors: string[];
  stagePreferences?: string[];
  /** @minLength 20 */
  bio: string;
}