# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CapitalBridge API
 * OpenAPI spec version: 0.1.0
 */
import type { DealStatus } from "./dealStatus";

export interface Deal {
  id: string;
  companyId: string;
  companyName: string;
  industry: string;
  raiseAmount: number;
  committedAmount: number;
  status: DealStatus;
  openedAt: Date;
  closesAt?: Date | null;
}