# lib/api-zod/src/generated/types/investor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CapitalBridge API
 * OpenAPI spec version: 0.1.0
 */
import type { InvestorType } from "./investorType";

export interface Investor {
  id: string;
  name: string;
  type: InvestorType;
  tagline: string;
  location: string;
  avatarUrl: string;
  /** Assets under management in USD */
  aum: number;
  dealsCompleted: number;
  checkSizeMin: number;
  checkSizeMax: number;
  focusSectors: string[];
  stagePreferences: string[];
  verified: boolean;
  featured: boolean;
}