# lib/api-zod/src/generated/types/investorDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CapitalBridge API
 * OpenAPI spec version: 0.1.0
 */
import type { Investor } from "./investor";
import type { InvestorDetailNotableInvestmentsItem } from "./investorDetailNotableInvestmentsItem";
import type { InvestorDetailPartnersItem } from "./investorDetailPartnersItem";

export type InvestorDetail = Investor & {
  bio: string;
  thesis: string;
  notableInvestments: InvestorDetailNotableInvestmentsItem[];
  partners: InvestorDetailPartnersItem[];
};