# lib/api-zod/src/generated/types/platformSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CapitalBridge API
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformSummaryCapitalByIndustryItem } from "./platformSummaryCapitalByIndustryItem";
import type { PlatformSummaryInvestorsByTypeItem } from "./platformSummaryInvestorsByTypeItem";

export interface PlatformSummary {
  /** USD raised across the platform */
  totalCapitalRaised: number;
  companiesListed: number;
  activeInvestors: number;
  dealsClosedYtd: number;
  averageDealSize: number;
  investorsByType: PlatformSummaryInvestorsByTypeItem[];
  capitalByIndustry: PlatformSummaryCapitalByIndustryItem[];
}